+++ docs/changelog.rst
@@ -9,6 +9,9 @@ In Development
 Breaking Changes
   * Completely changed the way that settings are provided. Rather than being specified in a separate python file that is then imported, behavior is now set through environment variables.
 
+Features
+  * :issue:`379`: Add endpoint to retrieve/update a user's Know Me subscription through Apple.
+
 
 ******
 v1.4.1